desktop: refactor media file handling to be in one file
nfci
